requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
evaluate
tqdm
tensorflow==2.10.1
plotly
ipykernel
openpyxl
nbformat
xgboost
lightgbm
dill
catboost
reportlab
sklearn_evaluation
django
-e .